requirements.txt
flask
flask-sqlalchemy
werkzeug
requests
markdown


postgrest-py
gotrue
realtime
storage3


gunicorn